requirements.txt
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.1
openai>=1.52.0
pytest>=8.3.2
ruff>=0.6.9
